# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TEST",
        "title": "API",
        "contact": {},
        "version": "1.7"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/events/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint adds a new event.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Add an event",
                "parameters": [
                    {
                        "description": "Event details to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.AddEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.AddEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    }
                }
            }
        },
        "/events/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint deletes an event by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Delete an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID to delete",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    }
                }
            }
        },
        "/events/edit": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint edits an existing event.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Edit an event",
                "parameters": [
                    {
                        "description": "Event details to edit",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.EditEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.EditEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    }
                }
            }
        },
        "/events/get": {
            "get": {
                "description": "This endpoint retrieves an event by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Get an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID to retrieve",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.GetEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    }
                }
            }
        },
        "/events/getall": {
            "get": {
                "description": "This endpoint retrieves all events with pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Get all events",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.GetAllEventsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    }
                }
            }
        },
        "/events/search": {
            "get": {
                "description": "This endpoint searches events by query with pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Search events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.GetAllEventsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "olympy_api-gateway_genproto_event_service.AddEventRequest": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Event"
                }
            }
        },
        "olympy_api-gateway_genproto_event_service.AddEventResponse": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Event"
                }
            }
        },
        "olympy_api-gateway_genproto_event_service.EditEventRequest": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Event"
                }
            }
        },
        "olympy_api-gateway_genproto_event_service.EditEventResponse": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Event"
                }
            }
        },
        "olympy_api-gateway_genproto_event_service.Event": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sport_type": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "olympy_api-gateway_genproto_event_service.GetAllEventsResponse": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Event"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "olympy_api-gateway_genproto_event_service.GetEventResponse": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/olympy_api-gateway_genproto_event_service.Event"
                }
            }
        },
        "olympy_api-gateway_genproto_event_service.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}